refactor(web): split ActiveWorkoutPanel.tsx — extract item card, checklist, badge, helpers
